+++ packages/fastify/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @vercel/fastify
 
+## 0.1.46
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.6.10
+
+## 0.1.45
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.6.9
+
 ## 0.1.44
 
 ### Patch Changes