+++ packages/koa/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @vercel/koa
 
+## 0.1.26
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.6.10
+
+## 0.1.25
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.6.9
+
 ## 0.1.24
 
 ### Patch Changes