+++ packages/python-analysis/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vercel/python-analysis
 
+## 0.8.1
+
+### Patch Changes
+
+- Add `prisma-client-py` support and the quirks system ([#15289](https://github.com/vercel/vercel/pull/15289))
+
 ## 0.8.0
 
 ### Minor Changes